pyproject.toml
[tool.poetry]
name = "calmlib"
version = "2.0.9"
description = "Calmlib"
authors = ["Reliable Magician <petr.b.lavrov@gmail.com>"]
readme = "README.md"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Development Status :: 4 - Beta",
    "Environment :: Other Environment",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

# todo: cleanup dependencies to speed up installation and resolving
packages = [
    { include = "calmlib", from = "." },
]


[tool.poetry.dependencies]
python = ">=3.11,<4.0"
loguru = ">=0.7.3"
# for parsing version
toml = ">=0.10"
python-dotenv = "*"
deprecated = "^1.2.17"
litellm = "^1.67.0.post1"


[tool.poetry.group.extras.dependencies]
# dependencies for extra features

# Main
## Pydantic
#pydantic = ">=2"            # Data validation and settings management
#pydantic-settings = ">=2"   # Pydantic settings management

# git ops
#PyGithub = "*"              # GitHub API client
#gitpython = "*"             # Git repository interface

# api
#fastapi = "*"               # High-performance API framework
#httpx = "*"                 # Async HTTP client

# cli
#click = "*"                 # CLI creation toolkit
#typer = "*"                 # Easy CLI building

# data, databases
pymongo = "*"                 # MongoDB client

# telegram bots
aiogram = "*"                 # Telegram bot framework
#mongoengine = "*"            # MongoDB ORM
#PyDrive = "*"                # Google Drive API client
protobuf = "*"               # Protocol buffers
#dropbox = "*"                # Dropbox API client

# utils
#pytz = "*"                  # Timezone definitions
#retry = "*"                 # Retry decorator
#nltk = "*"                  # Natural language toolkit
#easydict = ">=1.10"         # Easy access to dict keys - dotdict
pydub = ">=0.25"             # Audio manipulation
#apscheduler = ">=3"         # Scheduling library
pyperclip = "*"              # Copying to clipboard
aiolimiter = ">=1.1"         # Async rate limiting
deepl = ">=1.18"             # DeepL API client

# LLM, GPT, AI
#tiktoken = ">=0.7"            # Tokenization for OpenAI models
#openai = ">=1.45"             # OpenAI API client
langchain = "^0.3.14"          # Language model framework
langchain-openai = "^0.3.0"    # OpenAI integration for LangChain
langchain-community = "*"      # Community extensions for LangChain
langchain-anthropic = "^0.3.3" # Anthropic integration for LangChain
langchain-ollama = "^0.2.2"
langchain-google-genai = "^2.0.9"
langfuse = "*"                 # Language model analytics
#assemblyai = "*"              # Speech-to-text API
#whisperx = { git = "https://github.com/m-bain/whisperX.git", branch = "main" }  # Advanced Whisper model

# Misc
#ipykernel = "*"             # IPython kernel
#ipywidgets = "*"            # Interactive widgets for Jupyter
#joblib = "*"                # Lightweight pipelining
#paramiko = "*"              # SSH2 protocol library
#scp = "*"                   # Secure copy
#pymorphy2 = "*"             # Morphological analyzer

[tool.poetry.group.test.dependencies]
pytest = ">=6"
pytest-mongodb = "*"
vulture = "^2.14"
pytest-cov = "^6.1.1"
pytest-asyncio = "^0.26.0"
pytest-mock = "^3.14.0"
pyupgrade = "^3.19.1"
pyright = "^1.1.399"
ruff = "^0.11.6"

[tool.poetry.group.docs.dependencies]
#mkdocs = ">=1.6"

[tool.poetry.group.dev.dependencies]
# dev / research utils
jupyter = ">=1.0.0"          # Interactive notebooks
#jupyter-nbextensions-configurator = "*"
#jupyter_contrib_nbextensions = "*"
black = ">=24.0.0"           # Code formatter
pre-commit = ">=3.8.0"       # Pre-commit hooks
tqdm = "*"                   # Progress bars

## data analysis, research
#pandas = ">=2.0.0"          # Data manipulation and analysis
#seaborn = "*"               # Statistical data visualization
#bokeh = "*"                 # Interactive visualizations
#matplotlib = "*"            # Plotting library
#numpy = "*"                 # Numerical computing

# legacy. Todo: Sort out
#defaultenv = { git = "https://github.com/bobuk/defaultenv.git", branch ="master", optional = true }
#bmmb = { git = "https://github.com/calmmage/bot_me_maybe.git", branch ="main", optional = true }
#gpt_kit = { git = "https://github.com/calmmage/gpt_kit.git", branch = "main", optional = true }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Home" = "https://github.com/calmmage/calmlib"